Prototype for buying and selling

+++ XCode/takestok/takestok/ViewControllers/ImagesCollectionViewController.m
@@ -32,7 +32,7 @@
     return cell;
 }
 
--(int)collectionView:(UICollectionView *)collectionView numberOfItemsInSection:(NSInteger)section{
+-(NSInteger)collectionView:(UICollectionView *)collectionView numberOfItemsInSection:(NSInteger)section{
     return _images.count;
 }
